Fix SUDT deposit
